stock-news-crawler/requirements.txt
feedparser==6.0.11
requests==2.31.0
pyyaml==6.0.1
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==5.1.0
schedule==1.2.1
python-dotenv==1.0.0
jinja2==3.1.6

